Version 3.31.21 (based on e01e35fc2f50785e221eae2a4da5f3de718afdaf)
Performance and stability improvements on all platforms.

Cr-Commit-Position: refs/heads/candidates@{#25295}

+++ src/heap/store-buffer.cc
@@ -519,7 +519,7 @@ void StoreBuffer::IteratePointersToNewSpace(ObjectSlotCallback slot_callback,
                 if (!has_only_tagged_fields) {
                   for (Address slot = start_address; slot < end_address;
                        slot += kPointerSize) {
-                    if (helper.IsTagged(slot - obj_address)) {
+                    if (helper.IsTagged(static_cast<int>(slot - obj_address))) {
                       // TODO(ishell): call this once for contiguous region
                       // of tagged fields.
                       FindPointersToNewSpaceInRegion(slot, slot + kPointerSize,
